requirements.txt
line-bot-sdk
flask
PyYAML
dvc
numpy
spacy
ujson
args
